possibilité d'ajouter des traits physiques + cheveux + système de recommandation + modification de la bdd pour que ça fonctionne

+++ src/Controller/MemberController.php
@@ -3,10 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\Member;
+use App\Entity\PhysicalTraits;
 use App\Form\MemberType;
 use App\Repository\MemberRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
@@ -16,8 +18,15 @@ use Symfony\Component\Routing\Annotation\Route;
 class MemberController extends AbstractController
 {
     #[Route('/', name: 'app_member_index', methods: ['GET'])]
-    public function index(MemberRepository $memberRepository): Response
+    public function index(MemberRepository $memberRepository, Security $security): Response
     {
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
+
         return $this->render('security/login.html.twig', [
             'members' => $memberRepository->findAll(),
         ]);
@@ -50,16 +59,33 @@ class MemberController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_member_show', methods: ['GET'])]
-    public function show(Member $member): Response
+    public function show(Member $member, EntityManagerInterface $entityManager, Security $security): Response
     {
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        $existingPhysicalTrait = $entityManager->getRepository(PhysicalTraits::class)->findOneBy(['member' => $user]);
+
         return $this->render('member/show.html.twig', [
             'member' => $member,
+            'physicalTrait' => $existingPhysicalTrait
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_member_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Member $member, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Member $member, EntityManagerInterface $entityManager, Security $security): Response
     {
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $form = $this->createForm(MemberType::class, $member);
         $form->handleRequest($request);
 
@@ -67,7 +93,7 @@ class MemberController extends AbstractController
 
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_member_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_member_show', ['id' => $member->getId()], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('member/edit.html.twig', [
@@ -77,8 +103,15 @@ class MemberController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_member_delete', methods: ['POST'])]
-    public function delete(Request $request, Member $member, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Member $member, EntityManagerInterface $entityManager, Security $security): Response
     {
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$member->getId(), $request->request->get('_token'))) {
             $entityManager->remove($member);
             $entityManager->flush();
